[dart2js] Skip js interop test on csp config
The newly added test includes a call to eval and doesn't work in csp
mode. Added the skip to both status files to keep them in sync and
avoid a new breakage when we start running the migrated version of
the test on that config.

Change-Id: I55e32dbca03c8aa49436c3e6457d40e491920966
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/156722
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>