Make NNBD and non-NNBD versions of undefined_getter_test.dart.
This is in preparation for working on external and abstract fields,
which will require introducing some NNBD-only tests to this file.

In the process, I discovered a flow analysis bug
(https://github.com/dart-lang/sdk/issues/42957), which I plan to fix
in a follow-up CL.  For now the test that provoked the bug is marked
as failing.

Change-Id: I0767d650cfaecb125b0cea27c0c7971d3112deff
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/157464
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>