External variable/field support: suppress "final not initialized" errors.
External final fields and variables don't need to be initialized (in
fact they can't be initialized), so we shouldn't report "uninitialized
field" errors for them.

This CL also adds some basic tests of the functionality of external
final fields and variables, making sure they don't introduce setters,
and that the getters they introduce are properly resolved and that
their types are properly respected by inheritance checks.

Change-Id: Ic60042b22b0636b7316119b663629c1c6ced6e2b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/158200
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>