Add a safety check to analyzer messages codegen.
Previously, if the user added an `index` field to a message but failed
to add an `analyzerCode` field, code generation would just silently
ignore the index.  This was error prone.  Changed to report
`invalidAnalyzerCode`, which is the same error we report if the
analyzer code is improperly formatted.

Change-Id: I37de06058cd1938af2c9ef506412dc7960986269
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/158402
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>