Run AssignmentDriverResolutionTest cases in null safety mode.
Previously, these tests ran only in legacy mode.  This CL moves the
tests to a mixin so that they can be run in both modes.

This is necessary because in follow-up CLs I will be changing the type
analysis rules for binary expressions in null safety mode (in
accordance with
https://github.com/dart-lang/language/blob/master/accepted/future-releases/nnbd/number-operation-typing.md),
and I want to make sure not to lose test coverage.

Bug: https://github.com/dart-lang/sdk/issues/42629
Change-Id: I44a443205e9b36c698b7c736b75f8701aed7f947
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/158600
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>