[dart2js] Fix nullability handling in `createFromStaticType`.
Previously, we were creating nullable typemasks and then stripping off
the nullability if non-nullability was specified. We should be assuming
non-nullability as the default and adding nullability in if required.

The former approach has the issue that typemasks for `Null` keep getting
converted to empty typemasks when we try and remove the nullability.

Change-Id: I58907c1c45c756f023d67043ce88eb82c4b1589c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/158371
Reviewed-by: Stephen Adams <sra@google.com>