[infra] Tweak timing of lkgr and google builders
The google builder needs to be green to advance the lkgr branch, but is
competing with other jobs for it's share of the time and therefore might
be starved. Because it's very fast (<1 min) it can have a higher
priority without causing much impact. This CL gives it the same priority
as tryjobs and an aggressive 5 min execution timeout (in case it
misbehaves in the future).

The lkgr builder's frequences is increased from once per hour to 4 times
per hour. This will make it more likely that the rolls to dev will
pickup a recent commit.

Change-Id: Ia7825ec56b1b97b4188fbce8380076a6cb64e5e2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/159900
Auto-Submit: Alexander Thomas <athom@google.com>
Reviewed-by: William Hesse <whesse@google.com>
Commit-Queue: William Hesse <whesse@google.com>