Analyzer: Replace the use of package:html with a custom XML parser.
From the dartdoc: Android Manifest files are written in XML. In order
to validate an Android Manifest file, however, we do not need to parse
or retain each element. This parser understands which elements are
relevant to manifest validation. This parser does not validate the XML,
and if it encounters an error while parsing, no exception is thrown.
Instead, a parse result with [ParseResult.error] is returned.

Change-Id: I1aeb79f098aab2b4ca9b0ff54b027d41744c65ae
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/160182
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>