[dart2js] Add AbstractValueDomain.isTruthy
In spec-mode, conditions are forced to be true/false if necessary by a
HBoolConversion check. The refined output of HBoolConversion can be a
constant, allowing a branch of an if-then-else to be deleted.

If production-mode (-O3), this check is omitted, making it possible
for `null` and other values to flow to the condition. Without the
refinement of the HBoolConversion, the abstract value might not even
be a constant. To enable a branch of the if-then-else to be deleted as
an optimization (i.e. without changing the behaviour), we need to be
able to predict the JavaScript interpretation of the condition.

Change-Id: I44200adb7270a3feb26ce110ad4930cfa127f8c3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/160402
Reviewed-by: Mayank Patke <fishythefish@google.com>
Commit-Queue: Stephen Adams <sra@google.com>