Add more tests for the data-driven rename fix
There is one test marked as failing because we haven't implemented
support for it yet (discussion welcome). Getting the other tests to pass
prompted enabling the fix for several additional error codes.

Also, I'm not convinced the organization of the tests is optimal. It was
fairly easy to see missing test cases for the deprecated API cases, but
I find it much harder when the tests are divided based on the error code
that is produced in the code to be fixed. Discussion of alternatives is
also welcome.

Change-Id: I51af96c72ca86b37aac59c769b42c2c93012ed80
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/161181
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>