Add more tests of the add-type-parameter change
I added several tests. I also renamed most of the tests to remove a
component that was redundant with the class name after my previous
reorganization of the tests. Unfortunately this will probably make it
fairly difficult to see what changed. No test coverage was lost, so it
might be easier to just look at the new version of the file.

Change-Id: I5a1c9ce6e062f9520631a9ec4a9cfc34141bf2c8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/161663
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>