[dart:io] Concurrently resolve IPv4 and IPv6 addresses on iOS.
Original issue: On iOS, The problem is that people on iOS can have their
IPv6 misconfigured so it doesn't work, but it's still turned on.
Additionally getaddrinfo on that platform doesn't handle that case well,
meaning that the IPv6 lookup takes a long time to time out, delaying the
legitimate response.

This fix will issue a concurrent lookup for IPv4 and IPv6 on iOS.
Whichever IP comes first will be used to try to build the connection.

Bug: https://github.com/dart-lang/sdk/issues/41451
Change-Id: I7cda9845f03cf435245bbfad493d7443629b05c3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/143883
Commit-Queue: Zichang Guo <zichangguo@google.com>
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>
Reviewed-by: Siva Annamalai <asiva@google.com>