Remove the defaultValue from the add-parameter change
I originally thought we'd need this to handle updating overrides, but I
now think we can get the deafult value from the overridden method and
hence don't need it in the data file. We can always re-add it when we
support the override case if I'm wrong.

Change-Id: I1016841d18badf9deeb92ad2294e5ffe4af94ef2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/162380
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>