Add a date field for transforms
This also contains a few minor clean-up changes to the parser, mostly
dealing with keys that should have been required but weren't.

I'm tempted to make the date a required field and then use it to sort
the transforms so that they are applied from oldest to newest. Right
now it's optional but provides a way for users to document the date
for their own record keeping. Thoughts appreciated.

Change-Id: If3a22bbc2dea30f760f868c1ca1bf41aaf4654c9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/162382
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>