Enable the add-missing-required-argument fix for missing required named parameters under null safety
Note that the fix is disabled when we can't suggest anything other than
`null` in an opted-in library and the parameter can't accept `null`. We
might want to relax that restriction, even though we'd be introducing an
error, if we think it's better to provide some fix rather than none in
this case. Doing so would violate a long-standing design constraint,
which is why I haven't already done it. Reducing the temptation to do so
is why I provided non-null defaults for so many other types.

Change-Id: I114a495ee2bf3e01f02a51b87cc8659048e56aa2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/162440
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>